+++ CHANGELOG.md
@@ -1,3 +1,12 @@
-## 0.0.1
+## 0.1.0
+
+* `h3.maxPolyfillSize` were added
+* `h3.polyfill` were added
+* `h3.h3ToGeoBoundary` were added
+* `h3.radsToDegs` were added
+* `h3.degsToRads` were added
+* `geojson2h3.h3ToFeature` were added
+* `geojson2h3.h3SetToFeatureCollection` were added
+## 0.1.0
 
 * First version